refactoring styles
