+++ src/Pages/LoginPage/LoginPage.jsx
@@ -1,0 +1,19 @@
+import Salutation from 'components/Salutation/Salutation';
+import LogInForm from 'components/LogInForm';
+import LogInFormWrapper from 'components/LogInForm/LogInFormWrapper';
+
+import { StyledPaper } from './LoginPage.styled';
+
+const LoginPage = () => {
+  return (
+    <StyledPaper elevation={0}>
+      <Salutation />
+
+      <LogInFormWrapper>
+        <LogInForm />
+      </LogInFormWrapper>
+    </StyledPaper>
+  );
+};
+
+export default LoginPage;
