+++ src/components/Navigation/Navigation.jsx
@@ -1,34 +1,18 @@
-// import { useDispatch } from 'react-redux';
-import { Box, Link, styled } from '@mui/material';
-
 import { NavLink } from 'react-router-dom';
+
+import { StyledLink, StyledBox } from './Navigation.styled';
 
 const Navigation = () => {
   return (
-    <Box sx={{ display: 'flex', gap: '15px', alignItems: 'center' }}>
-      <StyledLink component={NavLink} to="home" sx={{ my: 2 }}>
+    <StyledBox>
+      <StyledLink component={NavLink} to="home" my={2}>
         Home
       </StyledLink>
-      <StyledLink
-        to="contacts"
-        component={NavLink}
-        sx={{ my: 2, display: 'block' }}
-      >
+      <StyledLink to="contacts" component={NavLink} my={2}>
         Contacts
       </StyledLink>
-    </Box>
+    </StyledBox>
   );
 };
-
-const StyledLink = styled(Link)`
-  color: inherit;
-  text-decoration: none;
-  display: block;
-  text-transform: uppercase;
-  &.active {
-    font-weight: 600;
-    text-decoration: underline;
-  }
-`;
 
 export default Navigation;
