+++ src/components/RegisterForm/RegisterForm.styled.js
@@ -1,17 +1,6 @@
-import styled from 'styled-components';
-import { styled as muiStyled } from '@mui/material';
 import { Form, Field } from 'formik';
-import { Paper, Button } from '@mui/material';
+import { Button, styled } from '@mui/material';
 import { theme } from 'theme';
-
-export const StyledPaper = muiStyled(Paper)`
-  background-color: ${theme.colors.background};
-  padding: 15px;
-  border-radius: 10px;
-  flex-grow: 1;
-  width: 50%;
-  align-self: center;
-`;
 
 export const StyledForm = styled(Form)`
   display: flex;
@@ -21,10 +10,10 @@ export const StyledForm = styled(Form)`
   max-width: 600px;
 `;
 
-export const Label = styled.label`
-  font-size: 24px;
-  font-weight: 700;
-`;
+// export const Label = styled.label`
+//   font-size: 24px;
+//   font-weight: 700;
+// `;
 
 export const Input = styled(Field)`
 height: 2rem;
@@ -33,7 +22,7 @@ height: 2rem;
   margin-bottom: 15px;
 `;
 
-export const StyledButton = muiStyled(Button)`
+export const StyledButton = styled(Button)`
   width: 100%;
   font-size: 18px; 
   font-weight: 400;
