+++ src/components/Salutation/Salutation.styled.jsx
@@ -1,0 +1,20 @@
+import { Box, styled, Typography } from '@mui/material';
+
+export const StyledBox = styled(Box)`
+  width: 55%;
+`;
+
+export const StyledTitle = styled(Typography)`
+  text-align: start;
+  align-self: center;
+  font-weight: 600;
+  line-height: 2;
+`;
+export const StyledSpan = styled(Typography)`
+  font-weight: 600;
+`;
+
+export const StyledParagraph = styled(Typography)`
+  text-indent: 15px;
+  text-align: start;
+`;
