+++ src/components/UserMenu/UserMenu.styled.jsx
@@ -1,0 +1,22 @@
+import { Box, Button, styled } from '@mui/material';
+
+import { theme } from '../../theme';
+
+export const StyledBox = styled(Box)`
+  display: flex;
+  gap: 15px;
+  align-items: center;
+`;
+
+export const StyledButton = styled(Button)`
+  display: block;
+  color: inherit;
+  color: inherit;
+  border-color: ${theme.colors.accent};
+
+  :hover {
+    background-color: ${theme.colors.accent};
+    color: ${theme.colors.white};
+    border-color: ${theme.colors.accent};
+  }
+`;
